+++ protocol/scripts/deploys/mainnet/core-deploy/readme.md
@@ -10,10 +10,6 @@ npx hardhat run scripts/deploys/mainnet/20-deploy-ops-manager.ts --network mainn
 
 * Need to set the InstantExitQueue as the ExitQueue for TempleStaking - `TempleStaking#setExitQueue`
 * Need to set VaultProxy as a Faith Manager in the Faith Contract - `Faith#addManager`
-* Transfer ownership of JoiningFee to MSig
-* Transfer ownership of OpsManager to MSig
-* Transfer ownership of VaultProxy to MSig
-* Transfer ownership of InstantExitQueue to MSig
 * Ensure VaultProxy is funded (not strictly necessary at this step but must be done before launch)
 * Update `protocol/scripts/deploys/helpers.ts` with the new addresses for JoiningFee, OpsManagerLib, OpsManager, VaultProxy and InstantExitQueue
 * Run 21-deploy-30min-vaults.ts - this will call `OpsManager#createVault` with appropriate parameters to create a 1month vault, with 4 7 vault instances
